Manual unit test database creation and seeding
The in-memory SQLite database proved to be a real pain in the ass
when using more than 1 database connections. The solution was
simple: use a disk file.

But, that's not all that happened. The migration scripts and seed
data that was used to create and seed the database through Knex's
facilities were discarded for a manual load process, and here is
the clincher, with the same schema file that is used to create the
development databases and eventually, production database. We're
also loading the seed data used for development.

This all means:

- The unit test database problem with multiple connection has been
  resolved.
- We use the same schema and seed data for unit testing and
  development.